Reconstruct all API related to lr scheduler, unify dygraph and static
